+++ ray3DXR_v3/ray3DXR_v3/SpriteObject.cpp
@@ -208,7 +208,7 @@ BOOL CSpriteObject::InitMesh()
 	};
 
 
-	WORD Indices[] =
+	DWORD Indices[] =
 	{
 		0, 1, 2,
 		0, 2, 3
